requirements.txt
dotenv
Django
whitenoise
psycopg2-binary
psycopg2
django-cors-headers
djangorestframework
djangorestframework-gis
gunicorn
python-dateutil
requests